+++ gms-server/src/main/java/org/gms/server/life/MonsterListener.java
@@ -9,6 +9,6 @@ import org.gms.client.Character;
 public interface MonsterListener {
 
     void monsterKilled(int aniTime);
-    void monsterDamaged(Character from, int trueDmg);
-    void monsterHealed(int trueHeal);
+    void monsterDamaged(Character from, Long trueDmg);
+    void monsterHealed(Long trueHeal);
 }
